requirements/main.txt
databases[aiomysql] == 0.9.0
fastapi == 0.128.0
python-dotenv == 1.2.1
python-json-logger == 4.0.0
PyYAML == 6.0.3
redis == 7.1.0
uvicorn == 0.40.0
uvloop == 0.22.1; sys_platform != "win32"
